requirements.txt
flask
sqlalchemy
flask_sqlalchemy
pillow
colorthief
python-slugify
gunicorn
